# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Our database management service, visualize in https://petstore.swagger.io/",
    "title": "Arcana.",
    "version": "1.0.0"
  },
  "host": "localhost:8081",
  "basePath": "/",
  "paths": {
    "/status": {
      "get": {
        "description": "checks if the serivce is reachable",
        "operationId": "status",
        "responses": {
          "200": {
            "$ref": "#/responses/StatusOK"
          }
        }
      }
    },
    "/v1/user/{used_id}/event": {
      "post": {
        "tags": [
          "event"
        ],
        "summary": "Creates a new event.",
        "operationId": "EventCreate",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "Identifier generated by the storage TODO: remove",
            "name": "id",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Owner",
            "description": "Number of the owner of the event",
            "name": "owner",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "Desired event name",
            "name": "name",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Description",
            "description": "More detailed description of the event",
            "name": "description",
            "in": "query"
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/EventResponse"
          },
          "400": {
            "description": "Request",
            "$ref": "#/responses/Bad"
          },
          "500": {
            "description": "Server Error",
            "$ref": "#/responses/Internal"
          }
        }
      }
    }
  },
  "responses": {
    "EventResponse": {
      "description": "EventResponse description object response",
      "headers": {
        "description": {
          "type": "string",
          "description": "More detailed description of the event"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "Identifier generated by the storage"
        },
        "name": {
          "type": "string",
          "description": "Desired event name"
        },
        "owner": {
          "type": "integer",
          "format": "int64",
          "description": "Number of the owner of the event"
        }
      }
    }
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  }
}